chore: add support for localization in the plugin
The plugin.xml file has been modified to include a new resource bundle for localizing messages in the plugin. This allows for better support of different languages and improves the overall user experience. The resource bundle is named "messages.SolidBundle". This change enables the plugin to provide localized messages to users, making it more accessible and user-friendly.