ci: update Java setup in Github Actions workflows
Changed the Java setup in Github Actions workflows to use gradle/gradle-build-action v2.7.0. Previously, caching was employed for the setup but now switched to Gradle build action. This change is incorporated to achieve better build performance, increased capabilities, and flexibility. The amendments are applied in both 'release' and 'run-ui-tests' workflows to maintain uniformity in the setup process.