build: update Gradle version and improve Java setup
Updated the gradle version from 8.1.1 to 8.2.1 to initiate use of the latest version. This enhances the overall functionalities and features provided by the tools. Also, updated the Java setup in gradlew script to ensure 'java' command is in the path. Added newline at end of 'settings.gradle.kts' for consistency.